Extract PendingCounter to it's own file
